Add possibility to change selection of already selected variant attribute
